Navigation SafeArgs
Used Navigation SafeArgs from Compose Navigation. Destinations must be "@Serializable", so kotlin serialization was implemented. But because of new version of kotlin serialization library, "kapt" for current version of Room wasn't compatible, so moved from "kapt" to "ksp" - it was recomended by compiler any way as it is much faster in compilation.