iOS App module
Created iOS app - first screen is launching on iOS.
Had to switch to "kotlin.native.cacheKind=none" as there was an error during building iOS and AS proposed this solution.
Adjusted Koin to Compose Multiplatfor so now it runs also on iOS.
Had to drop version of navigation compose to 2.8.0-alpha12 as due to libs versions conflict there was an error on iOS:
Constructor 'CupertinoScrollDecayAnimationSpec.<init>' can not be called: No constructor found for symbol 'androidx.compose.animation.core.cupertino

Deleted unused libraries and code.