fix: Message not appearing on screen init
Problem:
When SharedFlow emitted an element before any collector was created (e.g. a screen reading the flow), then this element was omitted.

Solution:
Change SharedFlow to Channel(Channel.BUFFERED) - Buffered Channel will keep an element until it is collected and when it's collected then it disappears from the channel, so there won't be a repeated message on configuration change.